Build Onsen UI v2 from onsen.io branch
This branch contains the documentation for the latest release of Onsen
UI, plus any fixes to the documentation to the release. We cannot build
from master because it will contain unreleased changes, and we cannot
build from the latest release tag because it will not contain any
documentation fixes since the release, so this onsen.io branch is
necessary.